define contract for evaluations
- DetermineWinner and GetHeuristicValue should only return values from
the current player's perspective

+++ src/two-player-games-working/GameRules.cs
@@ -5,6 +5,6 @@ namespace two_player_games_working
     public abstract class GameRules<TState> where TState: IState
     {
         public abstract List<TState> Expand(TState state);
-        public abstract int? DetermineWinner(TState state);
+        public abstract float? DetermineWinner(TState state);
     }
 }
